% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyprinodon_variegatus}
\alias{cyprinodon_variegatus}
\title{Data: Species Profile - Cyprinodon variegatus}
\format{
An object of class \code{data.frame} with 24 rows and 6 columns.
}
\usage{
cyprinodon_variegatus
}
\description{
Species profile data for given species
}
\keyword{datasets}
