% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyprinus_carpio}
\alias{cyprinus_carpio}
\title{Data: Species Profile - Cyprinus carpio}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
cyprinus_carpio
}
\description{
Species profile data for given species
}
\keyword{datasets}
