% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danio_rerio}
\alias{danio_rerio}
\title{Data: Species Profile - Danio rerio}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
danio_rerio
}
\description{
Species profile data for given species
}
\keyword{datasets}
