% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalFunctions-Computation.R
\name{gaussQuadInt}
\alias{gaussQuadInt}
\alias{midpointInt}
\alias{gaussQuadSub}
\alias{quad2D}
\alias{ZToB}
\alias{BToZ}
\alias{TemplateToParameters}
\alias{TemplateToTempProfile}
\alias{GenerateSpeciesTemplate}
\alias{TemplateToProfile}
\title{Computation Functions - Ellner et al., 2016 with some modifications}
\usage{
gaussQuadInt(L, U, order = 7)

midpointInt(L, U, m)

gaussQuadSub(L, U, order = 7, intervals = 1, breaks = NULL)

quad2D(FUN, wts1, wts2, nodes1, nodes2)

ZToB(z, alloIntercept, alloSlope)

BToZ(b, alloIntercept, alloSlope)

TemplateToParameters(profileObject, daysToRep = 365)

TemplateToTempProfile(profileDataFile)

GenerateSpeciesTemplate(fileName = "Profile-template")

TemplateToProfile(profileDataFile)
}
\arguments{
\item{L}{Lower bound for quadrature [float]}

\item{U}{Upper bound for quadrature [float]}

\item{order}{Number of quadrature points [integer]}

\item{m}{Number of points in interval from L to U}

\item{intervals}{User can specify the number of subintervals}

\item{breaks}{The locations of breaks between subintervals}

\item{FUN}{Function of two variables to integrate}

\item{wts1}{Weights for first variable}

\item{wts2}{Weights for second variable}

\item{nodes1}{Nodes for first variable}

\item{nodes2}{Nodes for second variable}

\item{z}{Size in mm}

\item{alloIntercept}{Intercept parameter of allometric function}

\item{alloSlope}{Slope parameter of allometric function}

\item{b}{mass in g}

\item{profileObject}{A FishToxTranslator parameter template file, either a 'read.csv()' of a template .csv file, i.e. parameterTemplate=read.csv("specie_profile.csv") OR parameterTemplate=get("scientific_name") OR tempProfile object from TemplateToTempProfile()}

\item{daysToRep}{number of days to replicate template life history parameters for [default=365]}

\item{profileDataFile}{name of input .csv Profile-template. [string, .csv]}

\item{fileName}{name for output .csv file. Output will be "fileName.csv" with the .csv automatically appended to the fileName string supplied [string] [default="Profile-template"]}
}
\value{
a quadrature set of weights and nodes for computation
}
\description{
This set of functions are used in the numerical discretization routines for \code{\link{SimulateModel}}
}
\section{Functions}{
\itemize{
\item \code{gaussQuadInt()}: Gauss-Legendre quadrature on interval (L,U)

\item \code{midpointInt()}: Midpoint quadrature rule

\item \code{gaussQuadSub()}: Gaussian-Legendre quadrature on subintervals of (L,U).

\item \code{quad2D()}: Integrate a function FUN of 2 variables using specified weights and nodes in each variable. FUN must accept two vectors as arguments and return a vector of function values

\item \code{ZToB()}: Allometric scaling function ZToB takes (mm) to (g)

\item \code{BToZ()}: Allometric scaling function BToZ takes (g) to (mm)

\item \code{TemplateToParameters()}: Change parameter template to data frame for computation

\item \code{TemplateToTempProfile()}: Change template data file into a temporary species object for visualization

\item \code{GenerateSpeciesTemplate()}: Generate a species template .csv file for creating new species life histories and importing back

\item \code{TemplateToProfile()}: Creates a temporary species profile that can be used in visualization. This temporary profile is added to species_library object within the session

}}
