% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lepomis_macrochirus}
\alias{lepomis_macrochirus}
\title{Data: Species Profile - Lepomis macrochirus}
\format{
An object of class \code{data.frame} with 23 rows and 5 columns.
}
\usage{
lepomis_macrochirus
}
\description{
Species profile data for given species
}
\keyword{datasets}
