% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oryzias_latipes}
\alias{oryzias_latipes}
\title{Data: Species Profile - Oryzias latipes}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
oryzias_latipes
}
\description{
Species profile data for given species
}
\keyword{datasets}
