% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_master}
\alias{parameters_master}
\title{Data: Master Parameter List}
\format{
An object of class \code{data.frame} with 42 rows and 10 columns.
}
\usage{
parameters_master
}
\description{
This includes the list and description of all parameters currently used in the Fish Toxicity Translator
}
\keyword{parameters_master}
