% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pimephales_promelas}
\alias{pimephales_promelas}
\title{Data: Species Profile - Pimephales promelas}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\usage{
pimephales_promelas
}
\description{
Species profile data for given species
}
\keyword{datasets}
