% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_library}
\alias{species_library}
\title{Data: Species Library}
\format{
An object of class \code{data.frame} with 7 rows and 5 columns.
}
\usage{
species_library
}
\description{
A reference for all species included in the model.  Common name, scientific name, and reference for species profile
data built into the R package
}
\keyword{datasets}
