% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCEM.R
\name{TCEM}
\alias{TCEM}
\title{Threshold Concentration Effect Model}
\usage{
TCEM(exposure_concentrations, lc_conc, lc_percent)
}
\arguments{
\item{exposure_concentrations}{time series (one concentration for each day, 365 days in length) of chemical exposures that are used to test if lc_conc is above or below [vector]}

\item{lc_conc}{Lethal concentration for 'lc_percent' [float]}

\item{lc_percent}{Percentage lethality in population for 'lc_conc' [float]}
}
\value{
Time series of survival decrements for the associated 'exposure_concentrations' time series. [vector]
}
\description{
This simplified effects model uses an lethal concentration for a given percent of the population to determine threshold type effects.
Specifically, a concentration 'lc_conc' and an associated percent 'lc_percent' combined with a time series of daily exposures and
when the 'exposure_concentrations' exceed the 'lc_conc' value, there is an 'lc_percent' decrease in survival that day.
}
